+++ src/components/publisherContainer.vue
